Refactors register and confirmation hooks out
