+++ lib/hooks/auth/index.ts
@@ -1,5 +1,3 @@
-export { useConfirm } from "./useConfirm";
 export { useLogin } from "./useLogin";
-export { useRegister } from "./useRegister";
 export { useResetPassword } from "./useResetPassword";
 export { useVerify } from "./useVerify";
