+++ pages/signup/register.tsx
@@ -1,9 +1,8 @@
-import React, { ReactElement } from "react";
+import React, { ReactElement, useRef, useState } from "react";
 import { Formik } from "formik";
 import { TextInput, Label, Alert, ErrorListItem, Description } from "@components/forms";
 import { Button } from "@components/globals";
 import { useFlag } from "@lib/hooks";
-import { useRegister } from "@lib/hooks/auth";
 import { useTranslation } from "next-i18next";
 import { GetServerSideProps } from "next";
 import { serverSideTranslations } from "next-i18next/serverSideTranslations";
@@ -23,12 +22,51 @@ import { getServerSession } from "next-auth/next";
 import Link from "next/link";
 import Head from "next/head";
 import { ErrorStatus } from "@components/forms/Alert/Alert";
+import { FormikHelpers } from "formik";
+import { fetchWithCsrfToken } from "@lib/hooks/auth/fetchWithCsrfToken";
+import { hasError } from "@lib/hasError";
+import { useAuthErrors } from "@lib/hooks/auth/useAuthErrors";
+import { logMessage } from "@lib/logger";
 
 const Register = () => {
   const { isLoading, status: registrationOpen } = useFlag("accountRegistration");
-  const { username, password, needsConfirmation, register, authErrorsState, authErrorsReset } =
-    useRegister();
   const { t } = useTranslation(["signup", "common"]);
+  const username = useRef("");
+  const password = useRef("");
+  const [needsConfirmation, setNeedsConfirmation] = useState(false);
+  const [authErrorsState, { authErrorsReset, handleErrorById }] = useAuthErrors();
+
+  const register = async (
+    {
+      username,
+      password,
+      name,
+    }: {
+      username: string;
+      password: string;
+      name: string;
+    },
+    { setSubmitting }: FormikHelpers<{ username: string; password: string; name: string }>
+  ) => {
+    authErrorsReset();
+    try {
+      await fetchWithCsrfToken("/api/signup/register", {
+        username,
+        password,
+        name,
+      });
+      setNeedsConfirmation(true);
+    } catch (err) {
+      logMessage.error(err);
+      if (hasError("UsernameExistsException", err)) {
+        handleErrorById("UsernameExistsException");
+        return;
+      }
+      handleErrorById(t("InternalServiceException"));
+    } finally {
+      setSubmitting(false);
+    }
+  };
 
   const validationSchema = Yup.object().shape({
     name: Yup.string()
@@ -91,11 +129,16 @@ const Register = () => {
 
   if (needsConfirmation) {
     return (
-      <Confirmation
-        username={username.current}
-        password={password.current}
-        confirmationCallback={() => undefined}
-      />
+      <>
+        <Head>
+          <title>{t("signUpRegistration.title")}</title>
+        </Head>
+        <Confirmation
+          username={username.current}
+          password={password.current}
+          confirmationCallback={() => undefined}
+        />
+      </>
     );
   }
